Fixed User Profile storage handler
